pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hzl-robotics-esp32"
version = "0.1.0"
description = "High speed ESP32 IO framework for robotics"
authors = [
  { name="Haziel Campos", email="user.haziel@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pyserial>=3.5"
]

[project.urls]
"Homepage" = "https://github.com/tuusuario/hzl_robotics"

[tool.setuptools.packages.find]
where = ["src"]
